Stop gap fix for an issue with RANLUX seed
GSL_RANLUX does not update the seed with every call like ran2, so every config saves the same seed. So continuation runs use the same PRNG sequence as the initial run. Not great.

Solution? Save a PRN as the "seed" instead. So the continuation run isn't the same as the original. Breaks Markov chain though...