Added the hip files.
We're having a few issues with the complex numbers. In CUDA theres the __NVCC__
flag used to signify __NVCC__ has been called to compile the host or device code.
The __CUDACC__ flag is only defined when compiling the device code. We use this
distinciton to get the C code to call the C standard complex number library and the
CUDA code to call the Thurst Complex number class.

ROCm lacks this distinction and only has a __HIPCC__ flag. We need to work around
this.