Okay, lots to get through here but all towards the same goal.
Main thing is, there is NO race condition in the conjate gradient
routine, or any of the dslash routines.

-Moved the CPU up/down partitioning routine from the end of main into
su2hmc.c. This new routine will instead call the GPU version for us if
we compile with CUDA. Makes the code a little easier to follow I think.

-Lots of new cudaDeviceSynchronise statements added, as well as debug
dependent print statements and managed memory

-Got rid of a debug statement in Dslashd_f so that the line numbers in
the GPU and CPU both match

-New diagnostic for Up/Down partitioning

Rebasing so the doxygen code features are applied to the CUDA branch.
Couple of little conflicts on the way but nothing serious