Good news: We have working GPU code in both the debug mode (Unified Memory) and release mode (manual memory management) Bad news: There are a couple of issues with the optimisation
If we compile with the -G flag for CUDA, the device optimisations are
disabled and the code produces the correct result no matter what compile
path or flags we set.

If however, we compile without -G (or -G -dopt=on to turn on
optimisations) it no longer produces the correct result, even
at -O0. Next it would be good to see what optimisations are turned on
by default so we can turn them on and off one by one to track down the
issue.

Reordered the async for the momentum field