23 cuda race condition (#36)
* Fixed an issue in the Diagnostics routine. Gauge_force was still using
being given double precision gauge fields when it now uses single

* Just makefile tweaks. The CUDA one isn't very portable but compiles
locally at least

* Okay, lots to get through here but all towards the same goal.

Main thing is, there is NO race condition in the conjate gradient
routine, or any of the dslash routines.

-Moved the CPU up/down partitioning routine from the end of main into
su2hmc.c. This new routine will instead call the GPU version for us if
we compile with CUDA. Makes the code a little easier to follow I think.

-Lots of new cudaDeviceSynchronise statements added, as well as debug
dependent print statements and managed memory

-Got rid of a debug statement in Dslashd_f so that the line numbers in
the GPU and CPU both match

-New diagnostic for Up/Down partitioning

* Main fix is in main.c around line 460ish. We convert a complex array
from single to double precision, but the GPU was calling the code to
convert a Real array between precisions, so half the elements were
ignored. Currently testing to see if results work out correctly.

* Good news: We have working GPU code in both the debug mode (Unified
Memory) and release mode (manual memory management)
Bad news: There are a couple of issues with the optimisation

If we compile with the -G flag for CUDA, the device optimisations are
disabled and the code produces the correct result no matter what compile
path or flags we set.

If however, we compile without -G (or -G -dopt=on to turn on
optimisations) it no longer produces the correct result, even
at -O0. Next it would be good to see what optimisations are turned on
by default so we can turn them on and off one by one to track down the
issue.

* Diagnostics have been taken out of matrices.c and into their own file

* Cleaned up Congradq a little

* Fixed an issue in the Diagnostics routine. Gauge_force was still using
being given double precision gauge fields when it now uses single

* Just makefile tweaks. The CUDA one isn't very portable but compiles
locally at least

* Okay, lots to get through here but all towards the same goal.

Main thing is, there is NO race condition in the conjate gradient
routine, or any of the dslash routines.

-Moved the CPU up/down partitioning routine from the end of main into
su2hmc.c. This new routine will instead call the GPU version for us if
we compile with CUDA. Makes the code a little easier to follow I think.

-Lots of new cudaDeviceSynchronise statements added, as well as debug
dependent print statements and managed memory

-Got rid of a debug statement in Dslashd_f so that the line numbers in
the GPU and CPU both match

-New diagnostic for Up/Down partitioning

Rebasing so the doxygen code features are applied to the CUDA branch.
Couple of little conflicts on the way but nothing serious

* Main fix is in main.c around line 460ish. We convert a complex array
from single to double precision, but the GPU was calling the code to
convert a Real array between precisions, so half the elements were
ignored. Currently testing to see if results work out correctly.

Lattice sizes changed between commits

* Good news: We have working GPU code in both the debug mode (Unified
Memory) and release mode (manual memory management)
Bad news: There are a couple of issues with the optimisation

If we compile with the -G flag for CUDA, the device optimisations are
disabled and the code produces the correct result no matter what compile
path or flags we set.

If however, we compile without -G (or -G -dopt=on to turn on
optimisations) it no longer produces the correct result, even
at -O0. Next it would be good to see what optimisations are turned on
by default so we can turn them on and off one by one to track down the
issue.

Reordered the async for the momentum field

* Diagnostics have been taken out of matrices.c and into their own file

Removed diagnostics from matrices.c

* Cleaned up Congradq a little

Moving the debug statements

* Forgot the header declaration for UpDownPartitioning

* Streams were an issue for cuforce. We cannot use separate streams for
each dirac index

* Minor debugging for non-zero density and diquark source

* Single GPU dense two colour QCD simulation works.