Gotta push (#44)
* Transpose routines now exist

* Extra comma

* Hdslash_f and Hdslashd_f are more like their GPU counterparts, and twice
as slow!?!

* Partially finished reducing u11t to ut[2] etc.

* Almost complete. But the Dslash_f operations are not working when the
gauge fields are changes to the single input

* Everything seems to be alright for now. Templates next

* Getting ready to break everything by going full SoA on the GPU

* Changes made, let's see if it works

* 31 prepare plaquette for clover (#38)

* Started doxygen of main.c

* Leaf indices are coming along

* Seems like the clover's almost there. Since I was commenting anyway a lot of this file now works with doxygen too

* Improved comments

* main.c marked up

* Congrad and Random have been marked up

* Fixed funky issue with comments

* Leaf tweaks. Want another sanity check before testing

* Documentation stuff

* C file documentation is going nicely.

* We've gotten into a rhythm now but need to go home

* Almost finished the first draft. Need a few bells and whistles like a
front page still but overall really happy with how it looks

* Turns out I never had the config file commited. Anyway, the dependency
graphs look unreal too!

* Update README.md

Slight tweaks to README

* Update README.md

More wee tweaks

* Auto doxygen

Automatically keep the doxygen page up to date

* Update main.yml

trying to get manual website build

* Update main.yml

* Update main.yml

* Update doxygen_config

* Update main.yml

* Update main.yml

* Update main.yml

* Update main.yml

* Update main.yml

* More doxygen

* More documenting tweaks

* Yet more documentation

* CPU code documentation basically done

* 28 doxygen (#35)

* More doxygen

* More documenting tweaks

* Yet more documentation

* CPU code documentation basically done

* Fixed missing maps due to missing fonts (locally at least!) and added a
css file for more fine-grained control of the output

* 23 cuda race condition (#36)

* Fixed an issue in the Diagnostics routine. Gauge_force was still using
being given double precision gauge fields when it now uses single

* Just makefile tweaks. The CUDA one isn't very portable but compiles
locally at least

* Okay, lots to get through here but all towards the same goal.

Main thing is, there is NO race condition in the conjate gradient
routine, or any of the dslash routines.

-Moved the CPU up/down partitioning routine from the end of main into
su2hmc.c. This new routine will instead call the GPU version for us if
we compile with CUDA. Makes the code a little easier to follow I think.

-Lots of new cudaDeviceSynchronise statements added, as well as debug
dependent print statements and managed memory

-Got rid of a debug statement in Dslashd_f so that the line numbers in
the GPU and CPU both match

-New diagnostic for Up/Down partitioning

* Main fix is in main.c around line 460ish. We convert a complex array
from single to double precision, but the GPU was calling the code to
convert a Real array between precisions, so half the elements were
ignored. Currently testing to see if results work out correctly.

* Good news: We have working GPU code in both the debug mode (Unified
Memory) and release mode (manual memory management)
Bad news: There are a couple of issues with the optimisation

If we compile with the -G flag for CUDA, the device optimisations are
disabled and the code produces the correct result no matter what compile
path or flags we set.

If however, we compile without -G (or -G -dopt=on to turn on
optimisations) it no longer produces the correct result, even
at -O0. Next it would be good to see what optimisations are turned on
by default so we can turn them on and off one by one to track down the
issue.

* Diagnostics have been taken out of matrices.c and into their own file

* Cleaned up Congradq a little

* Fixed an issue in the Diagnostics routine. Gauge_force was still using
being given double precision gauge fields when it now uses single

* Just makefile tweaks. The CUDA one isn't very portable but compiles
locally at least

* Okay, lots to get through here but all towards the same goal.

Main thing is, there is NO race condition in the conjate gradient
routine, or any of the dslash routines.

-Moved the CPU up/down partitioning routine from the end of main into
su2hmc.c. This new routine will instead call the GPU version for us if
we compile with CUDA. Makes the code a little easier to follow I think.

-Lots of new cudaDeviceSynchronise statements added, as well as debug
dependent print statements and managed memory

-Got rid of a debug statement in Dslashd_f so that the line numbers in
the GPU and CPU both match

-New diagnostic for Up/Down partitioning

Rebasing so the doxygen code features are applied to the CUDA branch.
Couple of little conflicts on the way but nothing serious

* Main fix is in main.c around line 460ish. We convert a complex array
from single to double precision, but the GPU was calling the code to
convert a Real array between precisions, so half the elements were
ignored. Currently testing to see if results work out correctly.

Lattice sizes changed between commits

* Good news: We have working GPU code in both the debug mode (Unified
Memory) and release mode (manual memory management)
Bad news: There are a couple of issues with the optimisation

If we compile with the -G flag for CUDA, the device optimisations are
disabled and the code produces the correct result no matter what compile
path or flags we set.

If however, we compile without -G (or -G -dopt=on to turn on
optimisations) it no longer produces the correct result, even
at -O0. Next it would be good to see what optimisations are turned on
by default so we can turn them on and off one by one to track down the
issue.

Reordered the async for the momentum field

* Diagnostics have been taken out of matrices.c and into their own file

Removed diagnostics from matrices.c

* Cleaned up Congradq a little

Moving the debug statements

* Forgot the header declaration for UpDownPartitioning

* Streams were an issue for cuforce. We cannot use separate streams for
each dirac index

* Minor debugging for non-zero density and diquark source

* Single GPU dense two colour QCD simulation works.

* Update README.md

* GCC makwfile for CUDA code works. Required a small change to how the
preprocessor handles SU2Plaq

* Removed Licence

* Create LICENSE

* Changes made, let's see if it works

* CPU test passed

* And the CUDA results are also acceptable

* An ugly but potentially useful approach
We add #ifdef __CLOVER__ statements to the function definitions and put
the clover stuff inside them. That way we don't need to make seperate
functions for the clover

* Minor fix after rebase

* Converting only the gauge to SoA everywhere. Works up until the first
force calculation.

* Forgot about transposes inside cuForce wrapper. Fixed now!

* Simulation on GPU is now 100% SoA with no swaps between AoS and SoA
during the simulation. Next up? Measurements

* Measure now accepts congrad. Just need to confirm output matches

* Polakov line works. So all measurements are now correct

* Made cuPolyakov prettier

* Profiler flags

* gsl-config added
CUDA makefiles now updated to use latest as NVHPC Path. You may need to
tweak this on your own machine

* Tweaks from AMD Hackathon made to the CUDA Code

* More index upgrades